chore: move PokeApi types and protocols to Common
