+++ PokeList/PokeList/Common/Models/PokeApiDataProtocol.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol PokeApiDataProtocol: Decodable {
+    var id: Int { get }
+}
