+++ PokeList/PokeList/Common/Models/PokeApiResponse.swift
@@ -7,11 +7,3 @@ struct PokeApiResponse: Decodable {
     let results: [PokeApiResultInfo]
 }
 
-struct PokeApiResultInfo: Decodable {
-    let name: String
-    let url: String
-}
-
-protocol PokeApiDataProtocol: Decodable {
-    var id: Int { get }
-}
