+++ PokeList/PokeList/Common/Models/PokeApiResultInfo.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct PokeApiResultInfo: Decodable {
+    let name: String
+    let url: String
+}
